src/pipeline_worker/requirements.txt
# Core
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.2
azure-identity==1.16.1
azure-storage-blob==12.22.0
azure-servicebus==7.12.2
httpx==0.27.2
tenacity==8.2.3
Pillow==10.2.0
supabase==2.9.0

# Background removal (rembg with ONNX)
rembg[gpu]==2.0.57

# Upscaling (Real-ESRGAN with PyTorch)
realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
